+++ Runtime/ZapparCameraBackground.cs
@@ -14,6 +14,9 @@ namespace Zappar
         private Matrix4x4 textureMatrix;
         private float[] textureMatElements = null;
         private Camera backgroundCamera = null;
+        private ZapparCamera mainCamera = null;
+
+        private float[] m_camerModel = null;
 
         public Texture2D GetCameraTexture => m_CamTexture;
         public Matrix4x4 GetTextureMatrix => textureMatrix;
@@ -28,6 +31,8 @@ namespace Zappar
             textureMatrix = new Matrix4x4();
             textureMatElements = new float[16];
             backgroundCamera = GetComponent<Camera>();
+            mainCamera = GetComponentInParent<ZapparCamera>();
+            m_camerModel = new float[] { 0, 0, 0, 0, 0, 0 };
         }
 
         void Point(float x, float y)
@@ -77,14 +82,16 @@ namespace Zappar
 #endif
         void Update()
         {
-            if (!m_Initialised || m_CameraMaterial == null)
+            if (m_CameraMaterial == null || mainCamera == null)
+                return;
+
+            if (!m_Initialised)
             {
-                if (Z.HasInitialized())
-                    m_Initialised = true;
+                m_Initialised = Z.HasInitialized() && mainCamera.CameraHasStarted;
                 return;
             }
 
-            backgroundCamera.projectionMatrix = Z.PipelineProjectionMatrix(ZapparCamera.Instance.GetPipeline, Screen.width, Screen.height);
+            backgroundCamera.projectionMatrix = Z.PipelineProjectionMatrix(ZapparCamera.Instance.GetPipeline, Screen.width, Screen.height, backgroundCamera.nearClipPlane, backgroundCamera.farClipPlane, ref m_camerModel);
 
             Z.PipelineCameraFrameTextureMatrix(ZapparCamera.Instance.GetPipeline, ref textureMatElements, Screen.width, Screen.height, ZapparCamera.Instance.IsMirrored);